% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_interactor_complex}
\alias{as_interactor_complex}
\title{as interactor complex}
\usage{
as_interactor_complex(x)
}
\description{
Converts a vector of N strings
to a 2 x (N ^ 2) data.frame that indicates 
a complex. I.e. all proteins interactor 
with all proteins in a directed manner.
}
\note{
duplicated (directed) entries are returned.
}
