% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_logfc.R
\name{calc_logfc}
\alias{calc_logfc}
\title{calculate log fold changes}
\usage{
calc_logfc(df)
}
\arguments{
\item{df}{a data.frame with bait and controls}
}
\description{
Calculates log fold change by substracting
a case column with a bait column. Assumes the data is
already in log2 intensities.
}
\note{
assummes log2(bait1), log2(control1), log2(bait2)..
}
\seealso{
Other processing: 
\code{\link{impute_lfq}()}
}
\author{
flassen
}
\concept{processing}
