% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_missing.R
\name{complete_missing}
\alias{complete_missing}
\title{Complete missing}
\usage{
complete_missing(
  df,
  id_var = "Prey",
  time_var = "Sheet",
  value_var = "LogFC",
  complete = 0,
  return_as = "long"
)
}
\arguments{
\item{df}{data.frame}

\item{id_var}{reshape2 id_var}

\item{time_var}{reshape2 time_var}

\item{value_var}{String. value column}

\item{complete}{what should the missing values be replaced with?}

\item{return_as}{string. Either "long" or "wide"}
}
\description{
Take a long data.frame and adds in missing values that would otherwise not be shown, when 
transforming the data to wide. This can primarily be used for plotting heatmaps.
}
\concept{plotting}
