% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{duplicated_interaction}
\alias{duplicated_interaction}
\title{duplicated interaction}
\usage{
duplicated_interaction(x)
}
\arguments{
\item{x}{a data.frame or matrix with at least two columns representing edges,
e.g 'Bait','Prey'}
}
\description{
Assuming an uni-directed graph, this function will return TRUE
for all entries that are duplicated.
}
\concept{graphs}
