% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_ms_data.R
\name{grep_ms_data}
\alias{grep_ms_data}
\title{get the right ms data columns}
\usage{
grep_ms_data(d, what = "Intensity", master, id = "Gene names")
}
\arguments{
\item{d}{a data.frame (typically an output from Mascott). This should ideally
contain colnames like this "Intensity 37", Intensity 38", "Intensity 39". Where
the first string indicates the name of column and the number indicates the well.}

\item{what}{name of column that is used for regex}

\item{master}{a master sheet that is used for subsetting. Should contain
at least one column 'well' that indicates the tube/sample of interest.}

\item{id}{what is the id column? Wil be keept alongside regexed columns}
}
\description{
a function used to handle Oxford Mass Spectromoetry intensity files.
}
\seealso{
Other processing: 
\code{\link{calc_logfc}()},
\code{\link{impute_by_col}()}
}
\concept{processing}
