% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_by_col.R
\name{impute_by_col}
\alias{impute_by_col}
\title{Impute values}
\usage{
impute_by_col(df, width = 0.3, shift = -1.8, seed = 4295, verbose = T)
}
\arguments{
\item{df}{a data.frame with numeric columns}

\item{width}{numeric. change the factor of the standard deviation.}

\item{shift}{numeric. Negative values will shift the median distribution downwards.}

\item{seed}{numeric. Random number seed to be used in the analysis.}
}
\value{
data.frame with missing values imputed.
}
\description{
Replacing missing values with randomly sampled values from normal distribution,
with width SD x width and down-shifted Median-Sd x shift compared to observed sample distribution.
This is building upon the assumption that missing values have arisen due to low expression that 
can't be quantified. Therfore, shifting the median to lower expression levels will provide a 
proxy of this. In contrast to, locf imputation, this ensures that the variance is not reduced
which would consequently impact the moderated t.test.
}
\note{
No down-shifting and stdwith of 0.5 do not simualte low abudant missing values.
down-shifting of 0.8 and stdwidth of 0.5 simulates low abundant missing values. 
down-shifting of 3.6 and stdwith of 0.5 results in (usually undesired) bi-modal distribution.
}
\references{
(Perseus, Tyanova et al. 2016)
}
\seealso{
Other processing: 
\code{\link{calc_logfc}()},
\code{\link{grep_ms_data}()}
}
\concept{processing}
