% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{index}
\alias{index}
\title{index}
\usage{
index(name, i, split = "_")
}
\arguments{
\item{name}{string or vector of strings.}

\item{i}{index}

\item{split}{how should the strings be splitted.}
}
\description{
index a vector by a splitted string.
}
