% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lower_triangle_duplicated}
\alias{lower_triangle_duplicated}
\title{lower triangle dupllicated}
\usage{
lower_triangle_duplicated(x, split = "\\\\.")
}
\arguments{
\item{x}{a vector of strings}

\item{split}{how the string should be splitted}
}
\description{
used to deal with a post-merge data.frame, 
and find unidirectional comparisons.
}
