% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_lfq.R
\name{norm_lfq}
\alias{norm_lfq}
\title{norm_lfq}
\usage{
norm_lfq(table = NULL, type = "median")
}
\arguments{
\item{table}{a table with numeric columns}

\item{type}{what function should be applied to columns}
}
\description{
Normalize a table of label or label free data. 
The function will calculate the \code{type} e.g. median of
all numeric columns and subtract the median from the 
intensity value.
}
\author{
flassen
}
