% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{read excel sheets}
\usage{
read_excel_sheets(path)
}
\arguments{
\item{path}{the path to the excel file.}
}
\value{
a list of tibbles
}
\description{
reads in excel sheets alongside their sheet-names as a list of tibbles.
}
\concept{excel}
